+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0 (2015-11-02)
+
+- Added: mdcss options are forwarded to theme
+- Added: `index` option from theme
+- Added: Moved some functionality from mdcss themes
+- Updated: Documentation and tests
+
 ## 1.0.1 (2015-10-30)
 
 - Added: Safely slash `destination` option